move theme support declarations to config namespace
